pyproject.toml
[project]
name = "proyecto-pln"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Alejandro Osornio", email = "aoxo.contact@gmail.com" }
]
dependencies = [
    "pip>=24.0",
    "streamlit>=1.34.0",
    "ipykernel>=6.29.4",
    "cleaner @ file:///home/ae/repos/proyecto-pln/target/wheels/cleaner-0.1.0-cp312-cp312-manylinux_2_34_x86_64.whl",
    "msgpack>=1.0.8",
    "scikit-learn>=1.4.2",
    "plotly-express>=0.4.1",
    "nbformat>=5.10.4",
    "transformers>=4.41.0",
    "torch>=2.3.0",
    "matplotlib>=3.9.0",
    "nltk>=3.8.1",
    "ipywidgets>=8.1.2",
    "spacy>=3.7.4",
    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl",
    "seaborn>=0.13.2",
    "tensorflow>=2.16.1",
    "contractions>=0.1.73",
    "wordcloud>=1.9.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
